requirements.txt
# MediSync AI — Unified Requirements
# Clinical NLP Safeguard + Agentic RAG Healthcare Navigator

# ─────────────────────────────────────────────────────────────
# Core Dependencies
# ─────────────────────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ─────────────────────────────────────────────────────────────
# LLM & Orchestration (LangChain + LangGraph)
# ─────────────────────────────────────────────────────────────
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-ollama>=1.0.0
langgraph>=0.2.0
langgraph-prebuilt>=0.1.0

# ─────────────────────────────────────────────────────────────
# Vector Database (ChromaDB)
# ─────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─────────────────────────────────────────────────────────────
# NLP & De-Identification (Presidio + spaCy)
# ─────────────────────────────────────────────────────────────
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.0

# ─────────────────────────────────────────────────────────────
# PDF Processing
# ─────────────────────────────────────────────────────────────
pypdf>=4.0.0

# ─────────────────────────────────────────────────────────────
# Data & Visualization
# ─────────────────────────────────────────────────────────────
pandas>=2.0.0
plotly>=5.22.0
faker>=25.0.0

# ─────────────────────────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────────────────────────
reportlab>=4.0.0